+++ internal/ast/ast.go
@@ -626,9 +626,15 @@ type SExportDefault struct {
 	Value       ExprOrStmt // May be a SFunction or SClass
 }
 
+type ExportStarAlias struct {
+	Loc  Loc
+	Name string
+}
+
 type SExportStar struct {
-	Item *ClauseItem
-	Path Path
+	NamespaceRef Ref
+	Alias        *ExportStarAlias
+	Path         Path
 }
 
 // This is an "export = value;" statement in TypeScript