+++ internal/bundler/linker.go
@@ -427,7 +427,18 @@ func (c *linkerContext) scanImportsAndExports() {
 		}
 	}
 
-	// Step 2: Resolve "export * from" statements. This must be done after we
+	// Step 2: Propagate CommonJS status for export star statements that are re-
+	// exports from a CommonJS module. Exports from a CommonJS module are not
+	// statically analyzable, so the export star must be evaluated at run time
+	// instead of at bundle time.
+	for _, sourceIndex := range c.reachableFiles {
+		if len(c.files[sourceIndex].ast.ExportStars) > 0 {
+			visited := make(map[uint32]bool)
+			c.isCommonJSDueToExportStar(sourceIndex, visited)
+		}
+	}
+
+	// Step 3: Resolve "export * from" statements. This must be done after we
 	// discover all modules that can be CommonJS because export stars are ignored
 	// for CommonJS modules.
 	for _, sourceIndex := range c.reachableFiles {
@@ -454,7 +465,7 @@ func (c *linkerContext) scanImportsAndExports() {
 		// Propagate exports for export star statements
 		if len(file.ast.ExportStars) > 0 {
 			visited := make(map[uint32]bool)
-			c.addExportsForExportStar(fileMeta.resolvedExports, uint32(sourceIndex), nil, visited)
+			c.addExportsForExportStar(fileMeta.resolvedExports, sourceIndex, nil, visited)
 		}
 
 		// Add an empty part for the namespace export that we can fill in later
@@ -479,7 +490,7 @@ func (c *linkerContext) scanImportsAndExports() {
 		file.ast.TopLevelSymbolToParts[file.ast.ExportsRef] = []uint32{nsExportPartIndex}
 	}
 
-	// Step 3: Match imports with exports. This must be done after we process all
+	// Step 4: Match imports with exports. This must be done after we process all
 	// export stars because imports can bind to export star re-exports.
 	for _, sourceIndex := range c.reachableFiles {
 		file := &c.files[sourceIndex]
@@ -510,7 +521,7 @@ func (c *linkerContext) scanImportsAndExports() {
 		}
 	}
 
-	// Step 4: Create namespace exports for every file. This is always necessary
+	// Step 5: Create namespace exports for every file. This is always necessary
 	// for CommonJS files, and is also necessary for other files if they are
 	// imported using an import star statement.
 	for _, sourceIndex := range c.reachableFiles {
@@ -548,7 +559,7 @@ func (c *linkerContext) scanImportsAndExports() {
 		c.createNamespaceExportForFile(uint32(sourceIndex))
 	}
 
-	// Step 5: Bind imports to exports. This adds non-local dependencies on the
+	// Step 6: Bind imports to exports. This adds non-local dependencies on the
 	// parts that declare the export to all parts that use the import.
 	for _, sourceIndex := range c.reachableFiles {
 		file := &c.files[sourceIndex]
@@ -834,6 +845,37 @@ func (c *linkerContext) matchImportsWithExportsForFile(sourceIndex uint32) {
 	}
 }
 
+func (c *linkerContext) isCommonJSDueToExportStar(sourceIndex uint32, visited map[uint32]bool) bool {
+	// Terminate the traversal now if this file is CommonJS
+	fileMeta := &c.fileMeta[sourceIndex]
+	if fileMeta.cjsStyleExports {
+		return true
+	}
+
+	// Avoid infinite loops due to cycles in the export star graph
+	if visited[sourceIndex] {
+		return false
+	}
+	visited[sourceIndex] = true
+
+	// Scan over the export star graph
+	file := &c.files[sourceIndex]
+	for _, path := range file.ast.ExportStars {
+		otherSourceIndex, isInternal := file.resolveImport(path)
+
+		// This file is CommonJS if the exported imports are from a file that is
+		// either CommonJS directly or transitively by itself having an export star
+		// from a CommonJS file.
+		if !isInternal || (otherSourceIndex != sourceIndex &&
+			c.isCommonJSDueToExportStar(otherSourceIndex, visited)) {
+			fileMeta.cjsStyleExports = true
+			return true
+		}
+	}
+
+	return false
+}
+
 func (c *linkerContext) addExportsForExportStar(
 	resolvedExports map[string]exportData,
 	sourceIndex uint32,
@@ -850,9 +892,7 @@ func (c *linkerContext) addExportsForExportStar(
 	for _, path := range file.ast.ExportStars {
 		otherSourceIndex, ok := file.resolveImport(path)
 		if !ok {
-			source := c.sources[sourceIndex]
-			c.addRangeError(source, source.RangeOfString(path.Loc),
-				"Wildcard exports are not supported for this module")
+			// This will be resolved at run time instead
 			continue
 		}
 
@@ -871,6 +911,7 @@ func (c *linkerContext) addExportsForExportStar(
 		// doing this we'd also have to rewrite any imports of these export star
 		// re-exports as property accesses off of a generated require() call.
 		if c.fileMeta[otherSourceIndex].cjsStyleExports {
+			// This will be resolved at run time instead
 			continue
 		}
 
@@ -1156,6 +1197,19 @@ func (c *linkerContext) includePart(sourceIndex uint32, partIndex uint32, entryP
 	if needsToModule {
 		c.includePartsForRuntimeSymbol("__toModule", entryPoint, distanceFromEntryPoint)
 	}
+
+	// If there's an ES6 export star statement of a non-ES6 module, then we're
+	// going to need the "__exportStar" symbol from the runtime
+	for _, path := range file.ast.ExportStars {
+		otherSourceIndex, isInternal := c.files[sourceIndex].resolveImport(path)
+
+		// Is this export star evaluated at run time?
+		if !isInternal || (otherSourceIndex != sourceIndex && c.fileMeta[otherSourceIndex].cjsStyleExports) {
+			c.includePartsForRuntimeSymbol("__exportStar", entryPoint, distanceFromEntryPoint)
+			file.ast.UsesExportsRef = true
+			break
+		}
+	}
 }
 
 func (c *linkerContext) computeChunks() []chunkMeta {
@@ -1382,23 +1436,69 @@ func (c *linkerContext) convertStmtsForChunk(sourceIndex uint32, stmtList *stmtL
 			}
 
 		case *ast.SExportStar:
-			if s.Item == nil {
+			if s.Alias == nil {
 				// "export * from 'path'"
 				if shouldStripExports {
-					// Remove re-export statements entirely
-					continue
+					otherSourceIndex, isInternal := c.files[sourceIndex].resolveImport(s.Path)
+
+					// Is this export star evaluated at run time?
+					if !isInternal && c.options.OutputFormat.KeepES6ImportExportSyntax() {
+						// Turn this statement into "import * as ns from 'path'"
+						stmt.Data = &ast.SImport{
+							NamespaceRef: s.NamespaceRef,
+							StarNameLoc:  &stmt.Loc,
+							Path:         s.Path,
+						}
+
+						// Prefix this module with "__exportStar(exports, ns)"
+						exportStarRef := c.files[ast.RuntimeSourceIndex].ast.ModuleScope.Members["__exportStar"]
+						stmtList.prefixStmts = append(stmtList.prefixStmts, ast.Stmt{
+							Loc: stmt.Loc,
+							Data: &ast.SExpr{ast.Expr{stmt.Loc, &ast.ECall{
+								Target: ast.Expr{stmt.Loc, &ast.EIdentifier{exportStarRef}},
+								Args: []ast.Expr{
+									ast.Expr{stmt.Loc, &ast.EIdentifier{c.files[sourceIndex].ast.ExportsRef}},
+									ast.Expr{stmt.Loc, &ast.EIdentifier{s.NamespaceRef}},
+								},
+							}}},
+						})
+
+						// Make sure these don't end up in a CommonJS wrapper
+						if shouldExtractES6StmtsForCJSWrap {
+							stmtList.es6StmtsForCJSWrap = append(stmtList.es6StmtsForCJSWrap, stmt)
+							continue
+						}
+					} else {
+						if !isInternal || (otherSourceIndex != sourceIndex && c.fileMeta[otherSourceIndex].cjsStyleExports) {
+							// Prefix this module with "__exportStar(exports, require(path))"
+							exportStarRef := c.files[ast.RuntimeSourceIndex].ast.ModuleScope.Members["__exportStar"]
+							stmtList.prefixStmts = append(stmtList.prefixStmts, ast.Stmt{
+								Loc: stmt.Loc,
+								Data: &ast.SExpr{ast.Expr{stmt.Loc, &ast.ECall{
+									Target: ast.Expr{stmt.Loc, &ast.EIdentifier{exportStarRef}},
+									Args: []ast.Expr{
+										ast.Expr{stmt.Loc, &ast.EIdentifier{c.files[sourceIndex].ast.ExportsRef}},
+										ast.Expr{s.Path.Loc, &ast.ERequire{Path: s.Path, IsES6Import: true}},
+									},
+								}}},
+							})
+						}
+
+						// Remove the export star statement
+						continue
+					}
 				}
 			} else {
 				// "export * as ns from 'path'"
-				if c.shouldRemoveImportExportStmt(sourceIndex, stmtList, partStmts, stmt.Loc, s.Item.Name.Ref, s.Path) {
+				if c.shouldRemoveImportExportStmt(sourceIndex, stmtList, partStmts, stmt.Loc, s.NamespaceRef, s.Path) {
 					continue
 				}
 
 				if shouldStripExports {
 					// Turn this statement into "import * as ns from 'path'"
 					stmt.Data = &ast.SImport{
-						NamespaceRef: s.Item.Name.Ref,
-						StarNameLoc:  &s.Item.Name.Loc,
+						NamespaceRef: s.NamespaceRef,
+						StarNameLoc:  &s.Alias.Loc,
 						Path:         s.Path,
 					}
 				}